sales_backend/sales_backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
cssselect2==0.8.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
html5lib==1.1
idna==3.10
lxml==6.0.1
oscrypto==1.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyJWT==2.10.1
pypdf==6.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
email-validator==2.1.1
dnspython==2.6.1
twilio==8.10.0
firebase-admin==6.2.0
pyotp==2.9.0
qrcode==7.4.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
